refactor: centralize manual repository helpers
